Fix trigger class selection
Up to now all trigger classes were used to check if we are in the active
data taking time. This information is then used to check the HV
conditions and to decide if to mask a ROC.
After the fix, only trigger classes where the TPC is active are used.